+++ frontend/package.json
@@ -18,10 +18,12 @@
   "dependencies": {
     "@material-ui/core": "^3.2.2",
     "@material-ui/icons": "^3.0.1",
+    "@material-ui/pickers": "3.0.0-beta.1",
     "@mdi/font": "^3.2.89",
     "@mdi/js": "^3.2.89",
     "@mdi/react": "^1.1.0",
     "@types/classnames": "^2.2.6",
+    "@y0c/react-datepicker": "^1.0.1",
     "apollo-boost": "^0.1.16",
     "classnames": "^2.2.6",
     "env-cmd": "^10.0.1",
@@ -32,6 +34,7 @@
     "msal": "^0.2.4",
     "react": "^16.8.4",
     "react-apollo": "^2.2.4",
+    "react-day-picker": "^7.3.2",
     "react-dom": "^16.8.4",
     "react-i18next": "^10.5.2",
     "react-markdown": "^4.0.6",
